# TRUST/docs/openapi.json
{"openapi":"3.1.0","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost","description":"Generated server url"}],"paths":{"/api/v1/imports":{"post":{"tags":["event-import-controller"],"operationId":"importFile","requestBody":{"content":{"multipart/form-data":{"schema":{"type":"object","properties":{"file":{"type":"string","format":"binary"}},"required":["file"]}}}},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{}}}}}}}},"/api/v1/evidence":{"post":{"tags":["evidence-controller"],"operationId":"upload","requestBody":{"content":{"multipart/form-data":{"schema":{"type":"object","properties":{"file":{"type":"string","format":"binary"}},"required":["file"]}}}},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{}}}}}}}},"/api/v1/events":{"get":{"tags":["event-controller"],"operationId":"list","parameters":[{"name":"q","in":"query","required":false,"schema":{"type":"string","default":""}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":30}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{}}}}}}},"post":{"tags":["event-controller"],"operationId":"submit","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EventInput"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/v1/events/{id}/verify":{"post":{"tags":["verification-controller"],"operationId":"verify","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{}}}}}}}},"/api/v1/events/{id}/retry":{"post":{"tags":["task-controller"],"operationId":"retry","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{}}}}}}}},"/api/v1/events/{id}/export":{"post":{"tags":["verification-controller"],"operationId":"export","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string","format":"binary"}}}}}}},"/api/v1/events/{id}/corrections":{"post":{"tags":["event-controller"],"operationId":"correct","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EventInput"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/v1/trace":{"get":{"tags":["trace-controller"],"operationId":"trace","parameters":[{"name":"kind","in":"query","required":true,"schema":{"type":"string"}},{"name":"value","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{}}}}}}}},"/api/v1/tasks":{"get":{"tags":["task-controller"],"operationId":"tasks","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"type":"object","additionalProperties":{}}}}}}}}},"/api/v1/status":{"get":{"tags":["operations-controller"],"operationId":"status","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{}}}}}}}},"/api/v1/me":{"get":{"tags":["identity-controller"],"operationId":"me","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{}}}}}}}},"/api/v1/evidence/{id}/download":{"get":{"tags":["evidence-controller"],"operationId":"download","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string","format":"byte"}}}}}}},"/api/v1/events/{id}":{"get":{"tags":["event-controller"],"operationId":"detail","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{}}}}}}}},"/api/v1/csrf":{"get":{"tags":["identity-controller"],"operationId":"csrf","parameters":[{"name":"token","in":"query","required":true,"schema":{"$ref":"#/components/schemas/CsrfToken"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}}}}},"/api/v1/audit":{"get":{"tags":["audit-controller"],"operationId":"audit","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"type":"object","additionalProperties":{}}}}}}}}}},"components":{"schemas":{"EventInput":{"type":"object","properties":{"sourceSystem":{"type":"string","minLength":1,"pattern":"[A-Za-z0-9._-]{1,80}"},"sourceEventId":{"type":"string","minLength":1,"pattern":"[A-Za-z0-9._-]{1,120}"},"eventType":{"type":"string","minLength":1,"pattern":"[A-Z_]{1,40}"},"businessObjectId":{"type":"string","maxLength":120,"minLength":0},"batchId":{"type":"string","maxLength":120,"minLength":0},"occurredAt":{"type":"string","format":"date-time"},"bundleIds":{"type":"array","items":{"type":"string","maxLength":120,"minLength":0},"maxItems":80,"minItems":0},"handoverId":{"type":"string","maxLength":120,"minLength":0},"relatedBatchIds":{"type":"array","items":{"type":"string","maxLength":120,"minLength":0},"maxItems":80,"minItems":0},"relatedEventRefs":{"type":"array","items":{"type":"string","pattern":"[A-Za-z0-9._-]{1,80}:[A-Za-z0-9._-]{1,120}"},"maxItems":80,"minItems":0},"quantity":{"type":"number","minimum":0},"unit":{"type":"string","maxLength":20,"minLength":0},"location":{"type":"string","maxLength":200,"minLength":0},"supplier":{"type":"string","maxLength":200,"minLength":0},"receiver":{"type":"string","maxLength":200,"minLength":0},"evidenceIds":{"type":"array","items":{"type":"string","pattern":"[a-f0-9-]{36}"},"maxItems":20,"minItems":0},"details":{"type":"object","additionalProperties":{"type":"string"}}},"required":["batchId","businessObjectId","eventType","occurredAt","sourceEventId","sourceSystem"]},"CsrfToken":{"type":"object","properties":{"parameterName":{"type":"string"},"token":{"type":"string"},"headerName":{"type":"string"}}}}}}